2020-01-21 16:50 updated

+++ db-phonebook/src/controller/MemberController.java
@@ -12,10 +12,10 @@ public class MemberController {
 	//controller
 
 	
-	public static void insertMember() {
+	public static void insertMember(int number) {
 		MemberVO member = new MemberVO();
 		
-		member=MemberView.printInsertMember();
+		member=MemberView.printInsertMember(number);
 		
 		MemberService.insertMember(member);
 	}
@@ -28,6 +28,9 @@ public class MemberController {
 		
 		memberList = MemberService.selectByName(2);
 		MemberView.printSelect(memberList);
+		if(memberList.size()==0) {
+			return;
+		}
 		MemberVO member = MemberView.SelectNumberToUpdate(memberList,2);
 		MemberService.deleteMember(member);
 		
@@ -37,9 +40,12 @@ public class MemberController {
 		
 		memberList = MemberService.selectByName(1);
 		MemberView.printSelect(memberList);
+		if(memberList.size()==0) {
+			return;
+		}
 		MemberVO member = MemberView.SelectNumberToUpdate(memberList,1);
 		MemberService.deleteMember(member);
-		insertMember();
+		insertMember(2);
 	}
 
 	public static void main(String[] args) {
@@ -48,7 +54,7 @@ public class MemberController {
 			
 			int selector = MemberView.viewIntro();
 			if(selector == 1) {
-				insertMember();
+				insertMember(1);
 			}else if(selector == 2){
 				selectAll();
 			}else if (selector == 3) {
@@ -59,7 +65,6 @@ public class MemberController {
 			}else if (selector == 5) {
 				// 5번일 경우 해당 while문을 빠져나간다.
 				// 반복문을 빠져 나가고 메인 블록을 닫는 중괄호로 가게 되면 출력되는 출력문
-				System.out.println("프로그램이 정상적으로 종료됩니다.");
 				return ;
 			}
 		}
